third_party/blink/web_tests/platform/win11-arm64/http/tests/devtools/network/load-resource-for-frontend-expected.txt

Test loading resource for frontend.


Running: testNetworkResourceNonCached
Loading resource from http://localhost:8080/devtools/network/resources/resource.php
Success: true
Logged headers: 
Content: "<br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>2</b><br />
<br />
<b>Warning</b>:  Undefined array key "wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>3</b><br />
<br />
<b>Warning</b>:  Undefined array key "tail_wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>4</b><br />
<br />
<b>Warning</b>:  Undefined array key "send" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>5</b><br />
<br />
<b>Warning</b>:  Undefined array key "size" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>6</b><br />
<br />
<b>Warning</b>:  Undefined array key "gzip" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>7</b><br />
<br />
<b>Warning</b>:  Undefined array key "jsdelay" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>8</b><br />
<br />
<b>Warning</b>:  Undefined array key "jscontent" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>9</b><br />
<br />
<b>Warning</b>:  Undefined array key "chunked" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>10</b><br />
<br />
<b>Warning</b>:  Undefined array key "random" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>11</b><br />
<br />
<b>Warning</b>:  Undefined array key "cached" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>12</b><br />
<br />
<b>Warning</b>:  Undefined array key "nosniff" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>13</b><br />
<br />
<b>Warning</b>:  Undefined array key "download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>14</b><br />
<br />
<b>Warning</b>:  Undefined array key "named_download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>15</b><br />
<br />
<b>Warning</b>:  Undefined array key "mime_type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>16</b><br />
<br />
<b>Warning</b>:  Undefined array key "body_pattern" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>17</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>44</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>45</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>46</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>53</b><br />
Hello world"

Running: testNetworkResourceCached
Loading resource from http://localhost:8080/devtools/network/resources/resource.php?cached=1
Success: true
Logged headers: 
Content: "<br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>2</b><br />
<br />
<b>Warning</b>:  Undefined array key "wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>3</b><br />
<br />
<b>Warning</b>:  Undefined array key "tail_wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>4</b><br />
<br />
<b>Warning</b>:  Undefined array key "send" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>5</b><br />
<br />
<b>Warning</b>:  Undefined array key "size" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>6</b><br />
<br />
<b>Warning</b>:  Undefined array key "gzip" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>7</b><br />
<br />
<b>Warning</b>:  Undefined array key "jsdelay" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>8</b><br />
<br />
<b>Warning</b>:  Undefined array key "jscontent" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>9</b><br />
<br />
<b>Warning</b>:  Undefined array key "chunked" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>10</b><br />
<br />
<b>Warning</b>:  Undefined array key "random" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>11</b><br />
<br />
<b>Warning</b>:  Undefined array key "nosniff" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>13</b><br />
<br />
<b>Warning</b>:  Undefined array key "download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>14</b><br />
<br />
<b>Warning</b>:  Undefined array key "named_download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>15</b><br />
<br />
<b>Warning</b>:  Undefined array key "mime_type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>16</b><br />
<br />
<b>Warning</b>:  Undefined array key "body_pattern" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>17</b><br />
<br />
<b>Warning</b>:  Undefined array key "HTTP_IF_MODIFIED_SINCE" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>24</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>39</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>40</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>41</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>42</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>53</b><br />
Hello world"

Running: testNetworkResourceIfModifiedSince
Loading resource from http://localhost:8080/devtools/network/resources/resource.php?cached=1
Success: true
Logged headers: 
Content: "<br />
<b>Warning</b>:  Undefined array key "type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>2</b><br />
<br />
<b>Warning</b>:  Undefined array key "wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>3</b><br />
<br />
<b>Warning</b>:  Undefined array key "tail_wait" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>4</b><br />
<br />
<b>Warning</b>:  Undefined array key "send" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>5</b><br />
<br />
<b>Warning</b>:  Undefined array key "size" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>6</b><br />
<br />
<b>Warning</b>:  Undefined array key "gzip" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>7</b><br />
<br />
<b>Warning</b>:  Undefined array key "jsdelay" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>8</b><br />
<br />
<b>Warning</b>:  Undefined array key "jscontent" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>9</b><br />
<br />
<b>Warning</b>:  Undefined array key "chunked" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>10</b><br />
<br />
<b>Warning</b>:  Undefined array key "random" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>11</b><br />
<br />
<b>Warning</b>:  Undefined array key "nosniff" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>13</b><br />
<br />
<b>Warning</b>:  Undefined array key "download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>14</b><br />
<br />
<b>Warning</b>:  Undefined array key "named_download" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>15</b><br />
<br />
<b>Warning</b>:  Undefined array key "mime_type" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>16</b><br />
<br />
<b>Warning</b>:  Undefined array key "body_pattern" in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>17</b><br />
<br />
<b>Warning</b>:  Cannot modify header information - headers already sent by (output started at C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php:2) in <b>C:\b\s\w\ir\third_party\blink\web_tests\http\tests\devtools\network\resources\resource.php</b> on line <b>25</b><br />
"

Running: testResourceFromIncorrectURL
Loading resource from resource.php
Success: false
Logged headers: 
Content: ""

Running: testResourceWithCookie
Loading resource from http://127.0.0.1:8000/devtools/network/resources/print-cookie.php
Success: true
Logged headers: cache-control, expires, pragma
Content: "Cookie value: TestCookieValue."
